refactor(benchmarks): move strict configuration models to owning loaders (#2004)
Move model families from strict_boundaries.py to their owning modules:
HarborJob* -> benchmark_job_models.py, Task*Section -> harbor_task_contract.py,
HeldoutRunPlan* -> heldout_plan_models.py, HeldoutManifest* ->
heldout_manifest_models.py. strict_boundaries.py retains only the shared
_StrictModel base and diagnostic helpers.
