+++ benchmarks/tooling/heldout_runner.py
@@ -15,7 +15,8 @@ from jsonschema import Draft202012Validator
 from benchmarks.tooling.command_runner import operator_environment, run_operator_command
 from benchmarks.tooling.errors import HarborSuiteError
 from benchmarks.tooling.harbor_suite import BENCHMARKS
-from benchmarks.tooling.strict_boundaries import HeldoutRunPlan, raise_strict_model
+from benchmarks.tooling.heldout_plan_models import HeldoutRunPlan
+from benchmarks.tooling.strict_boundaries import raise_strict_model
 
 CommandRunner = Callable[[list[str]], int]
